add CurpNode layer
